Internal change
PiperOrigin-RevId: 339131213
