+++ official/vision/image_classification/mnist_test.py
@@ -29,11 +29,14 @@ from official.utils.testing import integration
 from official.vision.image_classification import mnist_main
 
 
+mnist_main.define_mnist_flags()
+
+
 def eager_strategy_combinations():
   return combinations.combine(
       distribution=[
           strategy_combinations.default_strategy,
-          strategy_combinations.tpu_strategy,
+          strategy_combinations.cloud_tpu_strategy,
           strategy_combinations.one_device_strategy_gpu,
       ],
       mode="eager",
@@ -47,7 +50,6 @@ class KerasMnistTest(tf.test.TestCase, parameterized.TestCase):
   @classmethod
   def setUpClass(cls):  # pylint: disable=invalid-name
     super(KerasMnistTest, cls).setUpClass()
-    mnist_main.define_mnist_flags()
 
   def tearDown(self):
     super(KerasMnistTest, self).tearDown()
@@ -81,7 +83,7 @@ class KerasMnistTest(tf.test.TestCase, parameterized.TestCase):
     integration.run_synthetic(
         main=run,
         synth=False,
-        tmp_root=self.get_temp_dir(),
+        tmp_root=self.create_tempdir().full_path,
         extra_flags=extra_flags)
